fennec: mv resources to main module.
When importing multiple modules, a module couldn't see the resources declared
by the other modules so I moved the resources to the main module, prepending
them with "fennec_". Note that I didn't try hard to fix this.

The downside of having multiple modules for local projects is that you may need
to import all of the modules individually and we have to put all resources in
the main module. This may not be applicable if we moved these to a maven
repository, however.